chore(*): abstract code and fix runtime errors
