+++ projects/igniteui-angular/src/lib/date-common/picker-base.directive.ts
@@ -14,9 +14,14 @@ import { IgxInputGroupType, IGX_INPUT_GROUP_TYPE } from '../input-group/public_a
 import { OverlaySettings } from '../services/overlay/utilities';
 import { IgxPickerToggleComponent } from './picker-icons.common';
 import { PickerInteractionMode } from './types';
+import { getLocaleFirstDayOfWeek } from "@angular/common";
 
 @Directive()
 export abstract class PickerBaseDirective extends DisplayDensityBase implements IToggleView, EditorProvider, AfterViewInit, OnDestroy {
+    protected _locale;
+    protected _defaultLocaleFirstDayOfWeek: number;
+    protected _isWeekStartSet: boolean;
+
     /**
      * The editor's input mask.
      *
@@ -94,45 +99,47 @@ export abstract class PickerBaseDirective extends DisplayDensityBase implements 
     @Input()
     public disabled = false;
 
-    // /**
-    //  * Locale settings used for value formatting and calendar or time spinner.
-    //  *
-    //  * @remarks
-    //  * Uses Angular's `LOCALE_ID` by default. Affects both input mask and display format if those are not set.
-    //  * If a `locale` is set, it must be registered via `registerLocaleData`.
-    //  * Please refer to https://angular.io/guide/i18n#i18n-pipes.
-    //  * If it is not registered, `Intl` will be used for formatting.
-    //  *
-    //  * @example
-    //  * ```html
-    //  * <igx-date-picker locale="jp"></igx-date-picker>
-    //  * ```
-    //  */
-    // /**
-    //  * Gets the `locale` of the date-picker.
-    //  * Default value is `application's LOCALE_ID`.
-    //  */
-    // @Input()
-    // public get locale(): string {
-    //     return this._locale;
-    // }
-    //
-    // /**
-    //  * Sets the `locale` of the date-picker.
-    //  * Expects a valid BCP 47 language tag.
-    //  * Default value is `application's LOCALE_ID`.
-    //  */
-    // public set locale(value: string) {
-    //     debugger
-    //     this._locale = value;
-    //     try {
-    //         getLocaleFirstDayOfWeek(this._locale);
-    //     } catch (e) {
-    //         this._locale = this._localeId;
-    //     }
-    //
-    // }
+    /**
+     * Locale settings used for value formatting and calendar or time spinner.
+     *
+     * @remarks
+     * Uses Angular's `LOCALE_ID` by default. Affects both input mask and display format if those are not set.
+     * If a `locale` is set, it must be registered via `registerLocaleData`.
+     * Please refer to https://angular.io/guide/i18n#i18n-pipes.
+     * If it is not registered, `Intl` will be used for formatting.
+     *
+     * @example
+     * ```html
+     * <igx-date-picker locale="jp"></igx-date-picker>
+     * ```
+     */
+    /**
+     * Gets the `locale` of the date-picker.
+     * Default value is `application's LOCALE_ID`.
+     */
+    @Input()
+    public get locale(): string {
+        return this._locale;
+    }
 
+    /**
+     * Sets the `locale` of the date-picker.
+     * Expects a valid BCP 47 language tag.
+     * Default value is `application's LOCALE_ID`.
+     */
+    public set locale(value: string) {
+        this._locale = value;
+        try {
+            getLocaleFirstDayOfWeek(this._locale);
+        } catch (e) {
+            this._locale = this._localeId;
+        }
+
+        if (!this._isWeekStartSet) {
+            this._defaultLocaleFirstDayOfWeek = getLocaleFirstDayOfWeek(this._locale);
+        }
+
+    }
     /**
      * The container used for the pop-up element.
      *
